_block-nginx-access.sh: added (unused) blocking of '.git/config'
This is due to mostly one user on the Linode server network, using
unique spinup locations for every probe, and only one probe at a
time.

I believe they are looking for the _private_ origin server for a
given GIT repository, as a publicly accessible GIT repo would not
care if you accessed their repo.

Unless there is (also) a GIT 0-day we dont know about, in which
case a probe like this will alert you to an actual repo.