_block-ipv4-redo.sh: better log entry at boot/reboot + redo speedup
Before this change it took minutes to process this script, now that
there are almost 9000 block IPv4 address (not incl. ranges), which
is because TWO lists of 9000 must be cross-referenced, so this is
only ever going tyo take exponentially longer.

The speedup just stops 'grep' after the 1st match with 'grep -m1'.

NOTE: a faster "speedup" would be _not_ to cross-reference at all,
and instead just "apply the block" again, which results in 'ip'
outputting error messages. (I'll see when it gets past 10K blocks).